feat(ipc): implement capability generation and endpoint hardening
- Adds generation metadata to `cap_handle_t` to reject stale capability usage.
- Narrows broad capability rights (`CAP_RIGHT_SEND` -> `CAP_RIGHT_ENDPOINT_SEND`, `CAP_RIGHT_MAP` -> `CAP_RIGHT_MEMORY_MAP`, etc.).
- Explicit lifecycle state tracking added for both capabilities (`CAP_STATE_LIVE`, `FREE`, etc.) and endpoints (`IPC_ENDPOINT_STATE_READY`, `CLOSED`, etc.).
- Adds endpoint validation to verify capability state/generation during IPC send/receive paths.
- Introduces queue depth, drop counts, and timeout counters into the endpoint object structure.
- Refactors revocation paths to be cross-core safe using `CAP_STATE_REVOKING` to guard duplicate revokes.
- Updates and aligns kernel tests and host test cases to match the refactored capability contracts.

+++ kernel/include/cap_policy.h
@@ -7,6 +7,6 @@
 int cap_can_transfer(cap_type_t type);
 
 // Returns 1 if the rights requested to be transferred are valid for the given capability type.
-int cap_transfer_rights_valid(cap_type_t type, uint32_t transfer_rights);
+int cap_transfer_rights_valid(cap_type_t type, uint64_t transfer_rights);
 
 #endif // BHARAT_CAP_POLICY_H
